requirements.txt
Django
djangorestframework==3.11.0
coverage